<!-- Please use this line to close one or multiple issues when this pullrequest gets merged
You can add another line right under the first one:
resolves #1234
resolves #1235
-->

### This pullrequest changes

* Enable Faster-RCNN and RFCN object detection networks (from Caffe) with Inference Engine backend on CPU. There are fallbacks to default implementation at `DetectionOutput` and `Slice` layers.

  input size: 320x240

  | Model | OpenCV, CPU | OpenCV, GPU (fp32) | with IE, CPU | with IE, GPU (fp32) |
  |----|----|----|---|----|
  | VGG16 based | ~1.5s | 848ms | 357ms | |
  | ZF | 600ms | 320ms | 177ms | 1.5s |
  | ResNet-50 RFCN | 95ms | 134ms | 57ms | 115ms |

  CPU: Intel&reg; Core&trade; i7-6700K CPU @ 4.00GHz x 8
  GPU: Intel&reg; HD Graphics 530 (Skylake GT2)

* Enable Torch layers tests with Intel's Inference Engine backend.

**Merge with extra**: https://github.com/opencv/opencv_extra/pull/496

```
force_builders=Custom
docker_image:Custom=ubuntu-openvino:16.04
buildworker:Custom=linux-1
test_opencl:Custom=ON
```